manifest: Add LICENSE and copy it to checkout root
As required by the OSS office we need to have a LICENSE file at the root
directory of our source code. Instead of duplicating it in every gob
repository, we could use the `repo copyfile` trick that Chromium OS
uses to share a single copy of it across repositories.

The OSS office also requires that all source files that support comments
include the LICENSE header. The header in full.xml was auto-generated by
[addlicense](https://github.com/google/addlicense).

Change-Id: I753eab496e97af1c3db28240341c7dea47c0ce3c