manifest: Update repo names
Repos mirrored from Chromium are moved to the "mirrors/cros/" namespace.
On the other hand, COS authored repos now live in the "cos/" namespaces,
including the overlays and manifest.

Also, update README.md to reference the actual host.

The kernel repo hasn't been moved yet and is going to take a bit more
time. So it's commented out for now.

BUG=b:141581389
TEST=With a locally checked out manifest, `repo init -u manifest/.git &&
    repo sync && cros_sdk` works.

Change-Id: I93816ca5aa7c99db9a7f6b39a83c584c0919c7b4