Turn off cros_license_check for this repo
Now that we are using the Apache 2.0 license with the "Google LLC"
copyright header, we need to turn off the cros_license_check for this
repository, otherwise the repohook will complain. cros_license_check
will be replaced by the cos_license_check hook, that checks for the
Apache 2.0 license and the "Google LLC" copyright header.

BUG=b/149121748
TEST=tested when running `repo upload` for this cl

Change-Id: I85e4b20497d6b41caedfbb676f60d828f2c5ca0b