manifest: Add platform2 repository in external manifest.
In release-R77-12371.B, platform2 repository is not checked out
and therefore, ebuild files such as quipper depending on it tries
to clone and update it at run-time. This leads to transient build
failures. Having the repository present in manifest should avoid
these kind of problems.

Commit sha for platform2 is based on
https://cos.googlesource.com/third_party/overlays/chromiumos-overlay/+/refs/heads/release-R77-12371.B/chromeos-base/quipper/quipper-0.0.1-r2508.ebuild

BUG=b/167396326
TEST=repo init and repo sync

Change-Id: Id515f1af01d958a0f23397556ab8c6fae1a047f4