manifest: Use chromiumos-overlay from third_party.
Currently, R77 and R81 uses chromiumos-overlay from
https://cos.googlesource.com/mirrors/cros/chromiumos/overlays/chromiumos-overlay
where R85 and above uses from
https://cos.googlesource.com/third_party/overlays/chromiumos-overlay/

This creates if-else cases in the release notes tool while generating
changelog for chromiumos-overlay as the repo names are different.
As the commits currently used for R77 and R81 are present in
https://cos.googlesource.com/third_party/overlays/chromiumos-overlay/,
change the repo to
https://cos.googlesource.com/third_party/overlays/chromiumos-overlay/
so that all the branches are using the same repo and it avoids the
if-else condition for generating changelog.

BUG=b/172264418
TEST=repo init and sync

Change-Id: Iabc75a86133c0c8b274223e4bc9bf8f114b36c2f
Reviewed-on: https://cos-review.googlesource.com/c/cos/manifest/+/8282
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Vaibhav Rustagi <vaibhavrustagi@google.com>