manifest: Use forked chromite repository.
This CL is changing the chromite repository that will be checked-out and
used for builds. Apart from that it is also using the head of
release-R85-13310.B for chromite repository instead of using
a particular commit. This repository can be used for carrying patches on
top of ChromeOS chromite repo.

BUG=b/171561057
TEST=repo init and sync

Change-Id: I718a3e673ce805a0f0e6c4c3d0407d9842060787
Reviewed-on: https://cos-review.googlesource.com/c/cos/manifest/+/12890
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Vaibhav Rustagi <vaibhavrustagi@google.com>