manifest: Use forked chromite repository.
This CL is changing the chromite repository that will be checked-out and
used for builds. Apart from that it is also using the head of
release-R81-12871.B for chromite repository instead of using
a particular commit. This repository can be used for carrying patches on
top of ChromeOS chromite repo.

BUG=b/171561057
TEST=repo init and sync

Change-Id: Id99add59fe1bc46a45a6b097595e8e31f287c4ec
Reviewed-on: https://cos-review.googlesource.com/c/cos/manifest/+/12912
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Vaibhav Rustagi <vaibhavrustagi@google.com>