manifest: Use forked chromite repository.
This CL is changing the chromite repository that will be checked-out and
used for builds. Apart from that it is also using the head of
release-R77-12371.B for chromite repository instead of using
a particular commit. This repository can be used for carrying patches on
top of ChromeOS chromite repo.

BUG=b/171561057
TEST=repo init and sync

Change-Id: I931c7590a641a614e4a7a3faaa9ec43033e7f906
Reviewed-on: https://cos-review.googlesource.com/c/cos/manifest/+/12930
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Vaibhav Rustagi <vaibhavrustagi@google.com>